requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.1
ecdsa==0.19.1
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
python-dotenv==1.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlalchemy-continuum==1.4.2
sqlalchemy-utils==0.41.2
email-validator==2.2.0
python-multipart==0.0.20